data_colector/requirements.txt
asttokens==2.1.0
autopep8==1.7.0
backcall==0.2.0
backoff==2.2.1
black==22.10.0
boto3==1.26.13
botocore==1.29.13
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
idna==3.4
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.1
jmespath==1.0.1
jupyter_client==7.4.7
jupyter_core==5.0.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pyzmq==24.0.1
ratelimit==2.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
stack-data==0.6.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
